Fix compile error reported by Robert Kulagowski.
git-svn-id: http://svn.mythtv.org/svn/trunk@4270 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/decoderbase.cpp
@@ -178,7 +178,7 @@ bool DecoderBase::PosMapFromEnc(void)
         m_positionMap.push_back(e);
     }
     VERBOSE(VB_PLAYBACK, QString("Position map filled to: %1")
-            .arg(m_positionMap[m_positionMap.size()-1].index));
+            .arg((long int) m_positionMap[m_positionMap.size()-1].index));
 
     return true;
 }
